Moving changes from engine.cc to the rightful place optionsparser.cc
I had the changes wrongly in the uci protocol and that was pointed out to me by borg. So Moving the relevant changes to the right place in optionsparser.cc
