+++ src/engine.cc
@@ -30,8 +30,7 @@
 #include <algorithm>
 #include <cmath>
 #include <functional>
-#include <string>
-#include <vector>
+
 
 
 #include "mcts/search.h"
@@ -85,17 +84,6 @@ MoveList StringsToMovelist(const std::vector<std::string>& moves,
   }
   return result;
 }
-
-std::vector<std::string> SplitString(const std::string& str, char delimiter) {
-    std::vector<std::string> tokens;
-    std::string token;
-    std::istringstream tokenStream(str);
-    while (std::getline(tokenStream, token, delimiter)) {
-        tokens.push_back(token);
-    }
-    return tokens;
-}
-
 }  // namespace
 
 
@@ -372,33 +360,14 @@ void EngineLoop::CmdIsReady() {
   SendResponse("readyok");
 }
 
-void EngineLoop::CmdSetOption(const std::string& name, const std::string& value,
+void EngineLoop::CmdSetOption(const std::string& name, 
+                              const std::string& value,
                               const std::string& context) {
-    if (name.empty() || value.empty()) {
-        throw Exception("Both name and value must be provided");
-    } else {
-        // Split the names and values by semicolon.
-        std::vector<std::string> names = SplitString(name, ';');
-        std::vector<std::string> values = SplitString(value, ';');
-        // Check that the number of names and values match
-        if (names.size() != values.size()) {
-            throw Exception("Mismatched number of names and values");
-        }
-        // Set the UCI options for each name-value pair
-        for (size_t i = 0; i < names.size(); ++i) {
-            const std::string& name = names[i];
-            const std::string& value = values[i];
-            // Check that the name is valid
-            if (!options_.FindOptionByUciName(name)) {
-                throw Exception("Unknown option: " + name);
-            }
             // Set the UCI option
             options_.SetUciOption(name, value, context);
             // Set the log filename for the case it was set in UCI option.
             Logging::Get().SetFilename(
             options_.GetOptionsDict().Get<std::string>(kLogFileId));
-        }
-    }
 }
 
 void EngineLoop::CmdUciNewGame() { engine_.NewGame(); }
